backend/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pdf2image==1.17.0
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
websockets==15.0.1
